+++ deploy/kustomization.yaml
@@ -22,4 +22,4 @@ apiVersion: kustomize.config.k8s.io/v1beta1
 kind: Kustomization
 images:
 - name: opensourceway/omni-orchestrator
-  newTag: cae48c8096f236f600f40941c929dc3a34f62c46
+  newTag: 707c4ce86381392b590c0908d962681a93bbc353
